--- Hackerrank/30_Days_of_Code/Day-3.cpp
@@ -1,12 +1,0 @@
-#include <iostream>
-using namespace std;
-
-
-int main(){
-    int N;
-    cin >> N;
-    if(N%2 != 0 || (N%2 == 0) && N >= 6 && N <= 20){
-        cout << "Weird" << endl;
-    }else cout << "Not Weird" << endl;
-    return 0;
-}
